Misc changes

+++ std/Include/types.h
@@ -1,10 +1,10 @@
 #pragma once
 
-typedef unsigned long long u64;
-typedef signed long long s64;
+typedef unsigned long u64;
+typedef signed long s64;
 
-typedef unsigned long u32;
-typedef signed long s32;
+typedef unsigned int u32;
+typedef signed int s32;
 
 typedef unsigned short u16;
 typedef signed short s16;
